nn/requirements.txt
pyserial
tensorflow